# Refactor and optimize cdbpath_motion_for_join