--- a/src/test/regress/sql/join_gp.sql
+++ b/src/test/regress/sql/join_gp.sql
@@ -434,3 +434,67 @@ reset enable_nestloop;
 reset enable_material;
 reset enable_seqscan;
 reset enable_bitmapscan;
+
+-- Below test cases are for planner's cdbpath_motion_for_join, so we close
+-- ORCA temporarily.
+set optimizer = off;
+-- test outer join for general locus
+-- replicated table's locus is SegmentGeneral
+create table trep_join_gp (c1 int, c2 int) distributed replicated;
+-- hash distributed table's locus is Hash
+create table thash_join_gp (c1 int, c2 int) distributed by (c1);
+-- randomly distributed table's locus is Strewn
+create table trand_join_gp (c1 int, c2 int) distributed randomly;
+-- start_ignore
+create extension if not exists gp_debug_numsegments;
+select gp_debug_set_create_table_default_numsegments(1);
+-- end_ignore
+-- the following replicated table's numsegments is 1
+create table trep1_join_gp (c1 int, c2 int) distributed replicated;
+
+insert into trep_join_gp values (1, 1), (2, 2);
+insert into thash_join_gp values (1, 1), (2, 2);
+insert into trep1_join_gp values (1, 1), (2, 2);
+
+analyze trep_join_gp;
+analyze thash_join_gp;
+analyze trep1_join_gp;
+analyze trand_join_gp;
+
+-- This test is to check that: general left join segmentGeneral --> segmentGeneral
+-- And segmentGeneral join hash does not need motion.
+explain select * from generate_series(1, 5) g left join trep_join_gp on g = trep_join_gp.c1 join thash_join_gp on true;
+select * from generate_series(1, 5) g left join trep_join_gp on g = trep_join_gp.c1 join thash_join_gp on true;
+
+-- The following 4 tests are to check that general left join partition, we could redistribute the
+-- general-locus relation when the filter condition is suitable. If we can redistributed
+-- general-locus relation, we should not gather them to singleQE.
+explain select * from generate_series(1, 5) g left join thash_join_gp on g = thash_join_gp.c1;
+select * from generate_series(1, 5) g left join thash_join_gp on g = thash_join_gp.c1;
+
+explain select * from generate_series(1, 5) g left join thash_join_gp on g = thash_join_gp.c2;
+select * from generate_series(1, 5) g left join thash_join_gp on g = thash_join_gp.c2;
+
+explain select * from generate_series(1, 5) g left join trand_join_gp on g = trand_join_gp.c1;
+select * from generate_series(1, 5) g left join trand_join_gp on g = trand_join_gp.c1;
+
+explain select * from generate_series(1, 5) g full join trand_join_gp on g = trand_join_gp.c1;
+select * from generate_series(1, 5) g full join trand_join_gp on g = trand_join_gp.c1;
+
+-- The following 3 tests are to check that segmentGeneral left join partition
+-- we could redistribute the segment general-locus relation when the filter condition
+-- is suitable. If we can redistributed general-locus relation, we should not
+-- gather them to singleQE.
+explain select * from trep_join_gp left join thash_join_gp using (c1);
+select * from trep_join_gp left join thash_join_gp using (c1);
+
+explain select * from trep_join_gp left join trand_join_gp using (c1);
+select * from trep_join_gp left join trand_join_gp using (c1);
+
+explain select * from trep1_join_gp join thash_join_gp using (c1);
+select * from trep1_join_gp join thash_join_gp using (c1);
+
+drop table trep_join_gp;
+drop table thash_join_gp;
+drop table trand_join_gp;
+drop table trep1_join_gp;