… src/test/regress/expected/join_gp.out
… src/test/regress/expected/join_gp_optimizer.out
… src/test/regress/expected/partial_table.out