This PR refactor the function `cdbpath_motion_for_join` to make
it clear and generate better for some cases.

In distributed computing system, to gather distributed data into a
singleQE should always be the last choice. Previous code for general
and segmentgeneral, when they are not ok_to_replicate, will try to
gather other locus to singleQD. This commit improves this by firstly
trying to add redistributed motion.

The logic for the join result's locus (outer's locus is general):
```
  1. if outer is ok to replicated, then result's locus is the same
     as inner's locus
  2. if outer is not ok to replicated (like left join or wts cases)
     2.1 if inner's locus is hashed or hashOJ, we try to redistribute
         outer as the inner, if fails, make inner singleQE
     2.2 if inner's locus is strewn, we try to redistribute
         outer and inner, if fails, make inner singleQE
     2.3 just return the inner's locus, no motion is needed
```

The logic for the join results' locus (outer's locus is segmentgenral):
```
a. if both are SegmentGeneral:
     1. if both locus are equal, no motion needed, simply return
     2. For update cases. If resultrelation
        is SegmentGeneral, the update must execute
        on each segment of the resultrelation, if resultrelation's
        numsegments is larger, the only solution is to broadcast
        other
     3. no motion is needed, change both numsegments to common
 b. if only one of them is SegmentGeneral :
     1. consider update case, if resultrelation is SegmentGeneral,
         the only solution is to broadcast the other
     2. if other's locus is singleQE or entry, make SegmentGeneral
         to other's locus
     3. the remaining possibility of other's locus is partitioned
         3.1 if SegmentGeneral is not ok_to_replicate, try to
               add redistribute motion, if fails gather each to
               singleQE
         3.2 if SegmentGeneral's numsegments is larger, just return
               other's locus
         3.3 try to add redistribute motion, if fails, gather each
                to singleQE
```

Besides, Locus type Replicated can only be generated by join operation.
And in the function cdbpathlocus_join there is a rule:
    `<any locus type> join <Replicated> => any locus type`

Proof by contradiction, it shows that when code arrives
`cdbpath_motion_for_join`,
it is impossible that any of the two input paths' locus
is Replicated. So we add two asserts there.

We may generate more paths to let cost model choose and also we should learn from ORCA's result node skill to avoid motion. But I think this PR does make things clear and faster and the dev pipeline is green, we can do more improvement in future.

Some improved plans are shown below:

case 1: general join segmentgeneral

```sql
--old plan
explain (costs off) select * from  generate_series(1, 5) g
left join trep_join_gp on g = trep_join_gp.c1 right join
thash_join_gp on thash_join_gp.c1 < 10;
                            QUERY PLAN
------------------------------------------------------------------
 Nested Loop Left Join
   Join Filter: (thash_join_gp.c1 < 10)
   ->  Gather Motion 3:1  (slice1; segments: 3)
         ->  Seq Scan on thash_join_gp
   ->  Materialize
         ->  Hash Left Join
               Hash Cond: (g.g = trep_join_gp.c1)
               ->  Function Scan on generate_series g
               ->  Hash
                     ->  Gather Motion 1:1  (slice2; segments: 1)
                           ->  Seq Scan on trep_join_gp
--new plan
explain (costs off) select * from  generate_series(1, 5) g left join
trep_join_gp on g = trep_join_gp.c1 right join thash_join_gp
on thash_join_gp.c1 < 10;
                         QUERY PLAN
------------------------------------------------------------
 Gather Motion 3:1  (slice1; segments: 3)
   ->  Nested Loop Left Join
         Join Filter: (thash_join_gp.c1 < 10)
         ->  Seq Scan on thash_join_gp
         ->  Materialize
               ->  Hash Left Join
                     Hash Cond: (g.g = trep_join_gp.c1)
                     ->  Function Scan on generate_series g
                     ->  Hash
                           ->  Seq Scan on trep_join_gp
```

benchmark results: 

when  thash_join_gp contains 10000000 rows, the new plan is almost twice as faster as the old plan.

|plan| time|
|-----|----:|
|old|  10449.194ms|
|new|   6311.612ms|
 
------------------------------

case 2: general join hash

```sql
-- old plan
explain (costs off) select * from generate_series(1, 5) g
                                  left join thash_join_gp on
				  g = thash_join_gp.c1;
                      QUERY PLAN
------------------------------------------------------
Hash Left Join
   Hash Cond: (g.g = thash_join_gp.c1)
   ->  Function Scan on generate_series g
   ->  Hash
       ->  Gather Motion 3:1  (slice1; segments: 3)
           ->  Seq Scan on thash_join_gp
-- new plan
explain (costs off) select * from generate_series(1, 5) g left join
                                  thash_join_gp on g = thash_join_gp.c1;
                         QUERY PLAN
------------------------------------------------------------
Gather Motion 3:1  (slice2; segments: 3)
   ->  Hash Left Join
         Hash Cond: (g.g = thash_join_gp.c1)
         ->  Redistribute Motion 1:3  (slice1; segments: 1)
                     Hash Key: g.g
         ->  Function Scan on generate_series g
             ->  Hash
                 ->  Seq Scan on thash_join_gp
```

benchmark results: when  thash_join_gp contains 10000000 rows, the new plan is almost 3 times as faster as the old plan.

|plan|time|
|----|-----:|
|old| 6317.383 ms|
|new|2491.238ms|

------------------------------

case 3: general join strewn

```sql
-- old plan
explain (costs off) select * from generate_series(1, 5) g
                                  left join trand_join_gp on
				  g = trand_join_gp.c1;
                      QUERY PLAN
------------------------------------------------------
Hash Left Join
  Hash Cond: (g.g = trand_join_gp.c1)
  ->  Function Scan on generate_series g
  ->  Hash
      ->  Gather Motion 3:1  (slice1; segments: 3)
            ->  Seq Scan on trand_join_gp
-- new plan
explain (costs off) select * from generate_series(1, 5) g
                                  left join trand_join_gp
				  on g = trand_join_gp.c1;
                            QUERY PLAN
------------------------------------------------------------------
Gather Motion 3:1  (slice3; segments: 3)
  ->  Hash Left Join
        Hash Cond: (g.g = trand_join_gp.c1)
        ->  Redistribute Motion 1:3  (slice1; segments: 1)
                Hash Key: g.g
            ->  Function Scan on generate_series g
        ->  Hash
            ->  Redistribute Motion 3:3  (slice2; segments: 3)
                    Hash Key: trand_join_gp.c1
                    ->  Seq Scan on trand_join_gp
```

benchmark results: when  trand_join_gp contains 10000000 rows, the new plan is almost twice as faster as the old plan.

|plan|time|
|----|----:|
|old|6480.171ms|
|new|3155.293ms|

------------------------------

case 4: segmentgeneral left join hash

```sql
-- old plan
explain select * from trep left join thash using (c1);
                                          QUERY PLAN
---------------------------------------------------------------------------
Hash Right Join 
   Hash Cond: (thash.c1 = trep.c1)
   ->  Gather Motion 3:1  (slice1; segments: 3)  
         ->  Seq Scan on thash 
   ->  Hash
         ->  Gather Motion 1:1  (slice2; segments: 1)  
               ->  Seq Scan on trep

-- new plan
explain select * from trep left join thash using (c1);
                                               QUERY PLAN
---------------------------------------------------------------------------------
 Gather Motion 3:1  (slice2; segments: 3)
   ->  Hash Right Join
         Hash Cond: (thash.c1 = trep.c1)
         ->  Seq Scan on thash
         ->  Hash
               ->  Redistribute Motion 1:3  (slice1; segments: 1)
                     Hash Key: trep.c1
                     ->  Seq Scan on trep
```

benchmark results: when  trand_join_gp contains 100000000 rows, the new plan is almost 3 times as faster as the old plan.

|plan|time|
|----|----:|
|old|17274.039ms|
|new|5190.532ms|

## Here are some reminders before you submit the pull request
- [x] Add tests for the change
- [ ] Document changes
- [ ] Communicate in the mailing list if needed
- [x] Pass `make installcheck`
- [ ] Review a PR in return to support the community