@berlin-ab 2019-07-09
I was surprised to find plan changes in this PR after reading the title "Refactor ...". A refactor should change the internal structure without changing the behavior - given the planner is a black box.