@kainwen 2019-07-10
> I was surprised to find plan changes in this PR after reading the title "Refactor ...". A refactor should change the internal structure without changing the behavior - given the planner is a black box.

The intention at first is just to refactor the complicated code. During the process, I find some plans that can be optimized. So now I change the title to include the word "optimize".

> @kainwen Do you think this is ready to merge?

I think this is ready to merge. But there is no "+1" yet. 

Recently I am very busy with other jobs. May next week, I will try to invite some guys to review it deeply and push this job.