@kainwen 2019-07-22
Rebased with latest master.

I will run full dev-pipeline many times these days. If everything is OK, I will merge this into master.