@Stolb27 2023-06-27
@kainwen-zz, hello! Currently, we are working on support of CTE with DML operations over replicated tables. As a part of this work, we found out two cases when Locus Replicated may be created. One of them is Broadcast Motion path requested by join. The second is DML over replicated table. In case of equipping DML operation with RETURNING clause and wrapping into CTE, locus Replicated starts propagating up through plan. 

```
test=# explain with cte1 as (update r3 set a = 15 returning *) select * from cte1 join t1 using(a);
FATAL:  Unexpected internal error (assert.c:44)
DETAIL:  FailedAssertion("!(!((outer.locus).locustype == CdbLocusType_Replicated))", File: "cdbpath.c", Line: 1315)
server closed the connection unexpectedly
        This probably means the server terminated abnormally
        before or while processing the request.
The connection to the server was lost. Attempting reset: Failed.
```

So this statement is incorrect:

```
Besides, Locus type Replicated can only be generated by join operation.
And in the function cdbpathlocus_join there is a rule:
<any locus type> join <Replicated> => any locus type

Proof by contradiction, it shows that when code arrives
cdbpath_motion_for_join,
it is impossible that any of the two input paths' locus
is Replicated. So we add two asserts there.
```